Add finalizer
Added bits to echo in order to support the concept of a finalizer.  This is a func that will be called after everything else including the ErrorHandler.  This will allow more flexibility in how the request-response pipeline is executed.  For example, response data and error data can be stored in the context until the finalizer which can then determine how to output to the client.  This results in cleaner code when outputing data vs errors.